refactor: change widgets folders from voices_[name] to just [name] to be consistent
