+++ catalyst_voices/lib/widgets/app_bar/actions/notifications_indicator.dart
