+++ catalyst_voices/lib/widgets/app_bar/actions/unlock_button.dart
