+++ catalyst_voices/uikit_example/lib/examples/voices_navigation_example.dart
@@ -1,8 +1,8 @@
+import 'package:catalyst_voices/widgets/app_bar/actions/voices_app_bar_actions.dart';
+import 'package:catalyst_voices/widgets/app_bar/voices_app_bar.dart';
+import 'package:catalyst_voices/widgets/drawer/voices_drawer.dart';
 import 'package:catalyst_voices/widgets/menu/voices_expandable_list_tile.dart';
 import 'package:catalyst_voices/widgets/menu/voices_list_tile.dart';
-import 'package:catalyst_voices/widgets/voices_app_bar/actions/voices_app_bar_actions.dart';
-import 'package:catalyst_voices/widgets/voices_app_bar/voices_app_bar.dart';
-import 'package:catalyst_voices/widgets/voices_drawer/voices_drawer.dart';
 import 'package:catalyst_voices_assets/catalyst_voices_assets.dart';
 import 'package:flutter/material.dart';
